refactor a few more namespaces
